+++ datawaves/__init__.py
@@ -1,8 +1,10 @@
 from __future__ import absolute_import
 
 from datawaves.ethereum.condition import new_trace_conditions, new_log_conditions
+from datawaves.ethereum.providers import DatawavesABIProvider
 
 __all__ = [
     new_trace_conditions,
-    new_log_conditions
+    new_log_conditions,
+    DatawavesABIProvider
 ]
