+++ spark3/__init__.py
@@ -1,12 +1,11 @@
 from __future__ import absolute_import
 
 from spark3.main import Spark3, Transformer
-from spark3.providers import IContractABIProvider, EtherscanABIProvider, DatawavesABIProvider
+from spark3.providers import IContractABIProvider, EtherscanABIProvider
 
 __all__ = [
     Spark3,
     Transformer,
     IContractABIProvider,
     EtherscanABIProvider,
-    DatawavesABIProvider,
 ]
