+++ spark3/providers.py
@@ -33,13 +33,3 @@ class EtherscanABIProvider(IContractABIProvider):
             self.logger.error("Failed to get contract abi: %s:%s", data['status'], data['message'])
             raise FailToGetEtherscanABI(data['message'])
         return data['result']
-
-class DatawavesABIProvider(IContractABIProvider):
-    logger = logging.getLogger("spark3.providers.DatawavesProvider")
-
-    def get_contract_abi(self, contract_address: str) -> str:
-        r = requests.get("http://abi-cache.tellery-prod:3000/api/getabi", params={
-            'address': contract_address
-        })
-        r.raise_for_status()
-        return r.text
